backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Pillow==11.0.0
python-dotenv
Werkzeug==3.1.6
opencv-python==4.10.0.84
numpy>=2.0.0
scikit-learn==1.5.2
tensorflow>=2.20.0
matplotlib>=3.8.0
kagglehub==0.2.9
# ── AI Chatbot ─────────────────────────────────────────────────────────────────
groq==0.9.0
 
# ── Config / Utils ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1